fix: redone tests
