+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "goalkeeper-web",
   "private": true,
-  "version": "0.0.0",
+  "version": "0.0.1",
   "type": "module",
   "scripts": {
     "preinstall": "npx only-allow pnpm",
