+++ src/routes.tsx
@@ -1,0 +1,10 @@
+import Sample from './Sample'
+
+const routes = [
+  {
+    path: '/',
+    element: <Sample />,
+  },
+]
+
+export default routes
